+++ co_sched.h
@@ -1,8 +1,7 @@
 // A queue linked list node
-
 struct Qnode {
-    int priority;
-    int sleep_time;
+    short priority;
+    short sleep_time;
     void (*task_ptr);
     struct Qnode* next;
 };
